refactor: 상태 내리기 및 패딩 조정
