+++ snapsplit/src/features/trip/[tripId]/snap/SnapPage.tsx
@@ -2,26 +2,14 @@
 import { useState, useRef } from 'react';
 import TabSelector from '@/features/trip/[tripId]/snap/_components/TabSelector';
 import UploadButton from '@/features/trip/[tripId]/snap/_components/UploadButton';
-import SortBottomSheet from '@/features/trip/[tripId]/snap/_components/SortBottomSheet';
-import FilterBottomSheet from '@/features/trip/[tripId]/snap/_components/FilterBottomSheet';
 import BottomNavBar from '@/shared/components/BottomNavBar';
 import TripHeader from '@/features/trip/[tripId]/_components/TripHeader';
 import TripInfo from '@/features/trip/[tripId]/_components/TripInfo';
-import { FilterState } from '@/features/trip/[tripId]/snap/type';
-import { UploadedImage } from '@/features/trip/[tripId]/snap/type';
 import BaseTabView from '@/features/trip/[tripId]/snap/_components/tabView/BaseTabView';
 import FolderTabView from '@/features/trip/[tripId]/snap/_components/tabView/FolderTabView';
 import { ActiveTab } from '@/features/trip/[tripId]/snap/type';
 
 export default function SnapPage() {
-  const [sortOpen, setSortOpen] = useState(false);
-  const [filterOpen, setFilterOpen] = useState(false);
-  const [selectedSort, setSelectedSort] = useState('최신순');
-  const [filters, setFilters] = useState<FilterState>({
-    days: [],
-    people: [],
-    locations: [],
-  });
 
   const fileInputRef = useRef<HTMLInputElement>(null);
 
@@ -31,34 +19,6 @@ export default function SnapPage() {
     { countryId: 2, countryName: '파리' },
     { countryId: 3, countryName: '취리히' },
   ];
-
-  const testImages: UploadedImage[] = [
-    {
-      id: '1-jisu-london',
-      src: '/svg/1-jisu-london.png',
-      tags: {
-        days: [1],
-        people: ['지수'],
-        locations: ['런던']
-      }
-    },
-    {
-      id: '2-jisu-na-yeon-paris',
-      src: '/svg/2-jisu-na-yeon-paris.png',
-      tags: {
-        days: [2],
-        people: ['지수', '나경', '연수'],
-        locations: ['파리']
-      }
-    }
-  ];
-
-  const filteredImages = testImages.filter((img) => {
-    const matchDay = filters.days.length === 0 || filters.days.some((d) => img.tags.days.includes(d));
-    const matchPeople = filters.people.length === 0 || filters.people.some((p) => img.tags.people.includes(p));
-    const matchLocation = filters.locations.length === 0 || filters.locations.some((l) => img.tags.locations.includes(l));
-    return matchDay && matchPeople && matchLocation;
-  });
 
   const [activeTab, setActiveTab] = useState<ActiveTab>('전체');
 
@@ -78,13 +38,7 @@ export default function SnapPage() {
 
       {/* 컨텐츠 영역 */}
       {activeTab === '전체' ? (
-        <BaseTabView 
-          images={filteredImages}
-          selectedSort={selectedSort}
-          onSortOpen={() => setSortOpen(true)}
-          onFilterOpen={() => setFilterOpen(true)}
-          filters={filters}
-        />
+        <BaseTabView />
       ) : (
         <FolderTabView />
       )}
@@ -98,39 +52,7 @@ export default function SnapPage() {
           const file = e.target.files?.[0];
           if (file) alert(`파일 선택됨: ${file.name}`);
         }}
-      />
-      {sortOpen && (
-        <div className="fixed inset-0 z-100">
-          {/* 어두운 배경 */}
-          <div
-            className="absolute inset-0 bg-black opacity-50"
-            onClick={() => setSortOpen(false)}
-          />
-          {/* 정렬 bottom sheet */}
-          <SortBottomSheet
-            selectedSort={selectedSort}
-            onSelectSort={(opt) => setSelectedSort(opt)}
-            onClose={() => setSortOpen(false)}
-          />
-        </div>
-      )}
-      {filterOpen && (
-        <div className="fixed inset-0 z-100">
-          {/* 어두운 배경 */}
-          <div
-            className="absolute inset-0 bg-black opacity-50"
-            onClick={() => setFilterOpen(false)}
-          />
-          {/* 필터 bottom sheet */}
-          <FilterBottomSheet
-            filters={filters}
-            setFilters={setFilters}
-            onClose={() => setFilterOpen(false)}
-          />
-        </div>
-      )}
-      {/* 컨텐츠 영역 */}
-      
+      />      
 
       {/* 하단 네비게이션 */}
       <BottomNavBar />
