+++ snapsplit/src/features/trip/[tripId]/snap/_components/TabSelector.tsx
@@ -12,15 +12,15 @@ const tabs = ['전체', '폴더별'];
 export default function TabSelector({ activeTab, setActiveTab }: TabSelectorProps) {
 
   return (
-    <div className="flex justify-around mt-4 border-b border-grey-250">
+    <div className="flex h-11 border-b border-grey-250 text-body-1">
       {tabs.map((tab) => (
         <button
           key={tab}
           onClick={() => setActiveTab(tab as ActiveTab)}
-          className={`pb-2 transition-colors duration-200 w-full ${
+          className={`transition-colors duration-200 w-full ${
             activeTab === tab
-              ? 'font-semibold text-black border-b-2 border-black'
-              : 'text-grey-450'
+              ? 'border-b-1 border-black'
+              : ''
           }`}
         >
           {tab}
