+++ snapsplit/src/shared/components/BottomNavBar.tsx
@@ -15,15 +15,15 @@ export default function BottomNavBar() {
   ];
 
   return (
-    <nav className="fixed bottom-0 left-1/2 -translate-x-1/2 w-full max-w-[415px] lg:max-w-[360px] h-20 border-t border-grey-250 bg-white z-50">
-      <div className="flex justify-around w-full items-center py-2">
+    <nav className="fixed bottom-0 w-full h-15 max-w-[415px] lg:max-w-[360px] border-t border-grey-250 bg-white z-50">
+      <div className="flex w-full items-center py-2">
         {tabs.map((tab) => {
           const isActive = pathname === tab.path;
           return (
             <button
               key={tab.label}
               onClick={() => router.push(tab.path)}
-              className="flex flex-col items-center text-xs"
+              className="flex flex-col items-center w-full text-xs"
             >
               <div className="w-6 h-6 flex items-center justify-center">
                 <Image
